Add OnBoardingScreen.kt, moved MainActivity.kt to feature
