+++ app/src/main/AndroidManifest.xml
@@ -14,7 +14,7 @@
         android:theme="@style/Theme.DateMe"
         tools:targetApi="31">
         <activity
-            android:name="com.androiders.dateme.features.onboarding.ui.MainActivity"
+            android:name="com.androiders.dateme.features.MainActivity"
             android:exported="true"
             android:theme="@style/Theme.DateMe">
             <intent-filter>
