+++ app/src/main/java/com/androiders/dateme/features/MainActivity.kt
@@ -1,0 +1,30 @@
+package com.androiders.dateme.features
+
+import android.os.Bundle
+import androidx.activity.ComponentActivity
+import androidx.activity.compose.setContent
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.material.MaterialTheme
+import androidx.compose.material.Surface
+import androidx.compose.ui.Modifier
+import com.androiders.dateme.core.theme.DateMeTheme
+import com.androiders.dateme.features.onboarding.ui.screen.OnBoardingScreen
+import com.google.accompanist.pager.ExperimentalPagerApi
+
+class MainActivity : ComponentActivity() {
+    @OptIn(ExperimentalPagerApi::class)
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContent {
+            DateMeTheme {
+                // A surface container using the 'background' color from the theme
+                Surface(
+                    modifier = Modifier.fillMaxSize(),
+                    color = MaterialTheme.colors.background
+                ) {
+                    OnBoardingScreen()
+                }
+            }
+        }
+    }
+}
